tests: Move Helm Repo utils into test-utils
