+++ e2e/helm_test.go
@@ -3,174 +3,20 @@ package e2e
 import (
 	"github.com/go-git/go-git/v5"
 	"github.com/go-git/go-git/v5/plumbing/object"
-	"github.com/google/go-containerregistry/pkg/registry"
-	test_resources "github.com/kluctl/kluctl/v2/e2e/test-helm-chart"
-	"github.com/kluctl/kluctl/v2/pkg/utils"
+	test_utils "github.com/kluctl/kluctl/v2/e2e/test-utils"
 	"github.com/kluctl/kluctl/v2/pkg/utils/uo"
-	"github.com/kluctl/kluctl/v2/pkg/yaml"
 	"github.com/stretchr/testify/assert"
-	"helm.sh/helm/v3/pkg/action"
-	"helm.sh/helm/v3/pkg/cli"
-	"helm.sh/helm/v3/pkg/cli/values"
-	"helm.sh/helm/v3/pkg/getter"
-	"helm.sh/helm/v3/pkg/pusher"
 	registry2 "helm.sh/helm/v3/pkg/registry"
-	"helm.sh/helm/v3/pkg/repo"
-	"helm.sh/helm/v3/pkg/uploader"
-	"net/http"
-	"net/http/httptest"
-	"net/url"
-	"os"
 	"path/filepath"
 	"sort"
-	"strings"
 	"testing"
 )
 
-func createHelmPackage(t *testing.T, name string, version string) string {
-	tmpDir := t.TempDir()
-	err := utils.FsCopyDir(test_resources.HelmChartFS, ".", tmpDir)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	c, err := uo.FromFile(filepath.Join(tmpDir, "Chart.yaml"))
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	_ = c.SetNestedField(name, "name")
-	_ = c.SetNestedField(version, "version")
-
-	err = yaml.WriteYamlFile(filepath.Join(tmpDir, "Chart.yaml"), c)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	settings := cli.New()
-	client := action.NewPackage()
-	client.Destination = tmpDir
-	valueOpts := &values.Options{}
-	p := getter.All(settings)
-	vals, err := valueOpts.MergeValues(p)
-	retName, err := client.Run(tmpDir, vals)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	return retName
-}
-
-type repoChart struct {
-	chartName string
-	version   string
-}
-
-func createHelmRepo(t *testing.T, charts []repoChart, password string) string {
-	tmpDir := t.TempDir()
-
-	for _, c := range charts {
-		tgz := createHelmPackage(t, c.chartName, c.version)
-		_ = utils.CopyFile(tgz, filepath.Join(tmpDir, filepath.Base(tgz)))
-	}
-
-	fs := http.FileServer(http.FS(os.DirFS(tmpDir)))
-	s := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if password != "" {
-			_, p, ok := r.BasicAuth()
-			if !ok || p != password {
-				http.Error(w, "Auth header was incorrect", http.StatusUnauthorized)
-				return
-			}
-		}
-		fs.ServeHTTP(w, r)
-	}))
-
-	t.Cleanup(s.Close)
-
-	i, err := repo.IndexDirectory(tmpDir, s.URL)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	i.SortEntries()
-	err = i.WriteFile(filepath.Join(tmpDir, "index.yaml"), 0644)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	return s.URL
-}
-
-func createOciRepo(t *testing.T, charts []repoChart, password string) string {
-	tmpDir := t.TempDir()
-
-	ociRegistry := registry.New()
-	s := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if password != "" {
-			_, p, ok := r.BasicAuth()
-			if !ok {
-				w.Header().Add("WWW-Authenticate", "Basic")
-				http.Error(w, "Auth header was incorrect", http.StatusUnauthorized)
-				return
-			}
-			if !ok || p != password {
-				http.Error(w, "Auth header was incorrect", http.StatusUnauthorized)
-				return
-			}
-		}
-		ociRegistry.ServeHTTP(w, r)
-	}))
-
-	t.Cleanup(s.Close)
-
-	ociUrl := strings.ReplaceAll(s.URL, "http://", "oci://")
-	ociUrl2, _ := url.Parse(ociUrl)
-
-	var out strings.Builder
-	settings := cli.New()
-	c := uploader.ChartUploader{
-		Out:     &out,
-		Pushers: pusher.All(settings),
-		Options: []pusher.Option{},
-	}
-
-	var registryClient *registry2.Client
-	if password != "" {
-		var err error
-		registryClient, err = registry2.NewClient()
-		if err != nil {
-			t.Fatal(err)
-		}
-		err = registryClient.Login(ociUrl2.Host, registry2.LoginOptBasicAuth("test-user", password), registry2.LoginOptInsecure(true))
-		if err != nil {
-			t.Fatal(err)
-		}
-		c.Options = append(c.Options, pusher.WithRegistryClient(registryClient))
-	}
-
-	for _, chart := range charts {
-		tgz := createHelmPackage(t, chart.chartName, chart.version)
-		_ = utils.CopyFile(tgz, filepath.Join(tmpDir, filepath.Base(tgz)))
-
-		err := c.UploadTo(tgz, ociUrl)
-		if err != nil {
-			t.Fatal(err)
-		}
-	}
-
-	if registryClient != nil {
-		registryClient.Logout(ociUrl2.Host)
-	}
-
-	return ociUrl
-}
-
-func createHelmOrOciRepo(t *testing.T, charts []repoChart, oci bool, password string) string {
+func createHelmOrOciRepo(t *testing.T, charts []test_utils.RepoChart, oci bool, password string) string {
 	if oci {
-		return createOciRepo(t, charts, password)
+		return test_utils.CreateOciRepo(t, charts, password)
 	} else {
-		return createHelmRepo(t, charts, password)
+		return test_utils.CreateHelmRepo(t, charts, password)
 	}
 }
 
@@ -231,8 +77,8 @@ func testHelmPull(t *testing.T, tc testCase, prePull bool) {
 		password = "secret-password"
 	}
 
-	repoUrl := createHelmOrOciRepo(p.t, []repoChart{
-		{chartName: "test-chart1", version: "0.1.0"},
+	repoUrl := createHelmOrOciRepo(p.t, []test_utils.RepoChart{
+		{ChartName: "test-chart1", Version: "0.1.0"},
 	}, tc.oci, password)
 
 	p.updateTarget("test", nil)
@@ -316,9 +162,9 @@ func testHelmManualUpgrade(t *testing.T, oci bool) {
 
 	createNamespace(t, k, p.testSlug())
 
-	repoUrl := createHelmOrOciRepo(p.t, []repoChart{
-		{chartName: "test-chart1", version: "0.1.0"},
-		{chartName: "test-chart1", version: "0.2.0"},
+	repoUrl := createHelmOrOciRepo(p.t, []test_utils.RepoChart{
+		{ChartName: "test-chart1", Version: "0.1.0"},
+		{ChartName: "test-chart1", Version: "0.2.0"},
 	}, oci, "")
 
 	p.updateTarget("test", nil)
@@ -361,11 +207,11 @@ func testHelmUpdate(t *testing.T, oci bool, upgrade bool, commit bool) {
 
 	createNamespace(t, k, p.testSlug())
 
-	repoUrl := createHelmOrOciRepo(p.t, []repoChart{
-		{chartName: "test-chart1", version: "0.1.0"},
-		{chartName: "test-chart1", version: "0.2.0"},
-		{chartName: "test-chart2", version: "0.1.0"},
-		{chartName: "test-chart2", version: "0.3.0"},
+	repoUrl := createHelmOrOciRepo(p.t, []test_utils.RepoChart{
+		{ChartName: "test-chart1", Version: "0.1.0"},
+		{ChartName: "test-chart1", Version: "0.2.0"},
+		{ChartName: "test-chart2", Version: "0.1.0"},
+		{ChartName: "test-chart2", Version: "0.3.0"},
 	}, oci, "")
 
 	p.updateTarget("test", nil)
@@ -472,9 +318,9 @@ func TestHelmValues(t *testing.T) {
 
 	createNamespace(t, k, p.testSlug())
 
-	repoUrl := createHelmRepo(p.t, []repoChart{
-		{chartName: "test-chart1", version: "0.1.0"},
-		{chartName: "test-chart2", version: "0.1.0"},
+	repoUrl := test_utils.CreateHelmRepo(p.t, []test_utils.RepoChart{
+		{ChartName: "test-chart1", Version: "0.1.0"},
+		{ChartName: "test-chart2", Version: "0.1.0"},
 	}, "")
 
 	values1 := map[string]any{
@@ -537,9 +383,9 @@ func TestHelmTemplateChartYaml(t *testing.T) {
 	createNamespace(t, k, p.testSlug()+"-a")
 	createNamespace(t, k, p.testSlug()+"-b")
 
-	repoUrl := createHelmRepo(p.t, []repoChart{
-		{chartName: "test-chart1", version: "0.1.0"},
-		{chartName: "test-chart2", version: "0.1.0"},
+	repoUrl := test_utils.CreateHelmRepo(p.t, []test_utils.RepoChart{
+		{ChartName: "test-chart1", Version: "0.1.0"},
+		{ChartName: "test-chart2", Version: "0.1.0"},
 	}, "")
 
 	p.updateTarget("test", nil)
